requirements.txt
h11~=0.16.0
pip~=25.1.1
sniffio~=1.3.1
typing_extensions~=4.13.2
fastapi~=0.115.12
starlette~=0.46.2
pydantic~=2.11.5
pydantic_core~=2.33.2
websockets~=15.0.1
anyio~=4.9.0
click~=8.2.1
uvicorn~=0.34.2
httptools~=0.6.4
watchfiles~=1.0.5
annotated-types~=0.7.0
PyYAML~=6.0.2
numpy~=2.2.6
pymavlink~=2.4.47
typing_extensions~=4.13.2
pydantic_core~=2.33.2
pyserial~=3.5